feat: forgot password

+++ frontend/src/services/authService.js
@@ -1,100 +1,30 @@
 import api from "./api";
 
-
-
 const authService = {
-
-
-
     register: async (userData) => {
-
-
-        const response = await api.post(
-            "/auth/register",
-            userData
-        );
-
-
+        const response = await api.post("/auth/register", userData);
         return response.data;
-
     },
-
-
-
-
-
 
     login: async (credentials) => {
-
-
-        const response = await api.post(
-            "/auth/login",
-            credentials
-        );
-
-
+        const response = await api.post("/auth/login", credentials);
         return response.data;
-
     },
-
-
-
-
-
-
 
     changePassword: async (passwordData) => {
-
-
-        const response = await api.put(
-            "/users/change-password",
-            passwordData
-        );
-
-
+        const response = await api.put("/users/change-password", passwordData);
         return response.data;
-
     },
-
-
-
-
-
 
     getUsers: async () => {
-
-
-        const response = await api.get(
-            "/admin/users"
-        );
-
-
+        const response = await api.get("/admin/users");
         return response.data;
-
     },
 
-
-
-
-
-
     deleteUser: async (id) => {
-
-
-        const response = await api.delete(
-            `/admin/users/${id}`
-        );
-
-
+        const response = await api.delete(`/admin/users/${id}`);
         return response.data;
-
     }
-
-
-
-
 };
-
-
 
 export default authService;
